requirements.txt
Flask==2.2.5
requests==2.31.0
python-dotenv==0.21.1
